// src/models/Window.model.ts
export interface WindowI {
  id: string | number
  reference: string
  location: string
  cant: number
  series: string
  glass: string
  width: number
  height: number
  price: number
}

// ref(pin): "V1"
// location(pin): "Alcoba 1"
// series(pin): "Bella Sliding"
// width(pin): 1000
// height(pin): 1000
// cant(pin): 1
// glass(pin): "Templado 5mm"
// price(pin): 111.14838992874002